# api_server/openapi_server/openapi/openapi.yaml
---
openapi: 3.0.3
x-zally-ignore: [101, 104, 105]
info:
  contact:
    email: support@vwt.digital
    name: Digital Ambition Team
    url: https://vwt-digital.github.io/
  description: Endpoint to export car trips
  license:
    name: GNU GPLv3
    url: https://www.gnu.org/licenses/gpl.txt
  title: HR Car Usage export API
  version: 1.0.0
  x-audience: company-internal
  x-api-id: ced90f23-2e25-47f0-83f1-730638beb362
servers:
  - url: /
paths:
  /trips/export:
    get:
      description: Returns an export with a list of all car trips
      operationId: export_trips
      parameters:
        - $ref: '#/components/parameters/filterEndedAfterParam'
        - $ref: '#/components/parameters/filterEndedBeforeParam'
      responses:
        "200":
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: '#/components/schemas/TripsResponseFile'
            text/csv:
              schema:
                $ref: '#/components/schemas/TripsResponseFile'
            application/pdf:
              schema:
                $ref: '#/components/schemas/TripsResponseFile'
          description: Returns a list of car trips
        "204":
          description: No Content
        "400":
          content:
            application/problem+json:
              schema:
                $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
          description: No authorization token provided
        "403":
          content:
            application/problem+json:
              schema:
                $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
          description: Provided token doesn't have the required scope
      x-openapi-router-controller: openapi_server.controllers.default_controller
      security:
        - oauth2: [carusage.write]
components:
  parameters:
    filterEndedAfterParam:
      in: query
      name: ended_after
      required: false
      schema:
        type: string
        format: date-time
      description: Filter for trips that ended after a specific date
    filterEndedBeforeParam:
      in: query
      required: false
      name: ended_before
      schema:
        type: string
        format: date-time
      description: Filter for trips that ended before a specific date
  responses:
    "204":
      description: No Content
    "400":
      content:
        application/problem+json:
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
      description: Bad Request
    "401":
      content:
        application/problem+json:
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
      description: No authorization token provided
    "403":
      content:
        application/problem+json:
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
      description: Provided token doesn't have the required scope
    "404":
      content:
        application/problem+json:
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
      description: Resource Not Found
  schemas:
    TripFlatten:
      description: Information about a trip flattened
      example:
        license: AB-123-C
        started_at: 2020-01-01T12:00:00Z
        ended_at: 2020-01-01T13:00:00Z
        department: 123455
        function_name: Driver
        initial: An
        last_name: Niem
        prefix: O
        registration_number: 123456
        trip_kind: personal
        description: Driver drove to a supermarket
      properties:
        license:
          maxLength: 100
          type: string
        started_at:
          maxLength: 100
          type: string
          format: date-time
        ended_at:
          maxLength: 100
          type: string
          format: date-time
        department:
          type: integer
          format: int32
          minimum: 0
        function_name:
          maxLength: 100
          type: string
        initial:
          maxLength: 100
          type: string
        last_name:
          maxLength: 100
          type: string
        prefix:
          maxLength: 100
          type: string
        registration_number:
          type: integer
          format: int32
          minimum: 0
        trip_kind:
          default: work
          enum:
            - work
            - personal
          type: string
        description:
          maxLength: 250
          type: string
    TripsResponseFile:
      example:
        results:
          - id: 4e15f728-d9a3-4914-8e28-c0b184ae5f5f
            license: AB-123-C
            started_at: 2020-01-01T12:00:00Z
            ended_at: 2020-01-01T13:00:00Z
            department: 123455
            function_name: Driver
            initial: An
            last_name: Niem
            prefix: O
            registration_number: 123456
            trip_kind: personal
            description: Driver drove to a supermarket
      properties:
        results:
          items:
            $ref: '#/components/schemas/TripFlatten'
          type: array
      type: object
  securitySchemes:
    oauth2:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow.
      flows:
        implicit:
          authorizationUrl: https://azuread.url/2.0/authorize
          scopes:
            carusage.write: View access to Dynamic Data Manipulation API
      x-tokenInfoFunc: openapi_server.controllers.security_controller_.info_from_oAuth2
      x-scopeValidateFunc: connexion.decorators.security.validate_scope